refactor windws render code into multiple files
