+++ editor/EditorWindows.h
@@ -1,0 +1,32 @@
+// Copyright Contributors to the OpenVDB Project
+// SPDX-License-Identifier: Apache-2.0
+
+/*!
+    \file   nanovdb_editor/editor/EditorWindows.h
+
+    \author Andrew Reidmeyer, Petra Hapalova
+
+    \brief  Window rendering functions for the ImGui editor
+*/
+
+#pragma once
+
+namespace imgui_instance_user
+{
+struct Instance;
+
+// Window rendering functions
+void showSceneWindow(Instance* ptr);
+void showCameraViewWindow(Instance* ptr);
+void showPropertiesWindow(Instance* ptr);
+void showViewportSettingsWindow(Instance* ptr);
+void showRenderSettingsWindow(Instance* ptr);
+void showCompilerSettingsWindow(Instance* ptr);
+void showShaderParamsWindow(Instance* ptr);
+void showBenchmarkWindow(Instance* ptr);
+void showFileHeaderWindow(Instance* ptr);
+void showCodeEditorWindow(Instance* ptr);
+void showProfilerWindow(Instance* ptr, float delta_time);
+void showConsoleWindow(Instance* ptr);
+
+} // namespace imgui_instance_user
