+++ editor/ImguiInstance.cpp
@@ -14,14 +14,12 @@
 #include "imgui.h"
 #include "imgui/ImguiRenderer.h"
 #include "compute/ComputeShader.h"
+#include "CameraFrustum.h"
+#include "EditorWindows.h"
 #include "CodeEditor.h"
 #include "Console.h"
-#include "Node2Verify.h"
-#include "Profiler.h"
 #include "RenderSettingsHandler.h"
 #include "InstanceSettingsHandler.h"
-#include "ShaderMonitor.h"
-#include "ShaderCompileUtils.h"
 
 #include <ImGuiFileDialog.h>
 
@@ -35,25 +33,7 @@
 #    include <time.h>
 #endif
 
-#include <filesystem>
 #include <algorithm>
-#include <cmath>
-#include <limits>
-
-#ifndef M_PI_2
-#    define M_PI_2 1.57079632679489661923
-#endif
-
-#define IMGUI_CHECKBOX_SYNC(label, var)                                                                                \
-    {                                                                                                                  \
-        bool temp_bool = ((var) != PNANOVDB_FALSE);                                                                    \
-        if (ImGui::Checkbox((label), &temp_bool))                                                                      \
-        {                                                                                                              \
-            (var) = temp_bool ? PNANOVDB_TRUE : PNANOVDB_FALSE;                                                        \
-        }                                                                                                              \
-    }
-
-const float EPSILON = 1e-6f;
 
 PNANOVDB_INLINE void timestamp_capture(pnanovdb_uint64_t* ptr)
 {
@@ -122,226 +102,6 @@ void destroy(pnanovdb_imgui_instance_t* instance)
     delete ptr;
 }
 
-static const char* VIEWPORT_SETTINGS = "Viewport";
-static const char* RENDER_SETTINGS = "Render Settings";
-static const char* COMPILER_SETTINGS = "Compiler";
-static const char* PROFILER = "Profiler";
-static const char* CODE_EDITOR = "Shader Editor";
-static const char* CONSOLE = "Log";
-static const char* SHADER_PARAMS = "Params";
-static const char* BENCHMARK = "Benchmark";
-static const char* FILE_HEADER = "File Header";
-static const char* SCENE = "Scene";
-
-static const char* getGridTypeName(uint32_t gridType)
-{
-    switch (gridType)
-    {
-    case 0:
-        return "UNKNOWN";
-    case 1:
-        return "FLOAT";
-    case 2:
-        return "DOUBLE";
-    case 3:
-        return "INT16";
-    case 4:
-        return "INT32";
-    case 5:
-        return "INT64";
-    case 6:
-        return "VEC3F";
-    case 7:
-        return "VEC3D";
-    case 8:
-        return "MASK";
-    case 9:
-        return "HALF";
-    case 10:
-        return "UINT32";
-    case 11:
-        return "BOOLEAN";
-    case 12:
-        return "RGBA8";
-    case 13:
-        return "FP4";
-    case 14:
-        return "FP8";
-    case 15:
-        return "FP16";
-    case 16:
-        return "FPN";
-    case 17:
-        return "VEC4F";
-    case 18:
-        return "VEC4D";
-    case 19:
-        return "INDEX";
-    case 20:
-        return "ONINDEX";
-    case 21:
-        return "INDEXMASK";
-    case 22:
-        return "ONINDEXMASK";
-    case 23:
-        return "POINTINDEX";
-    case 24:
-        return "VEC3U8";
-    case 25:
-        return "VEC3U16";
-    case 26:
-        return "UINT8";
-    case 27:
-        return "NODE2";
-    default:
-        return "INVALID";
-    }
-}
-
-static const char* getGridClassName(uint32_t gridClass)
-{
-    switch (gridClass)
-    {
-    case 0:
-        return "UNKNOWN";
-    case 1:
-        return "LEVEL_SET";
-    case 2:
-        return "FOG_VOLUME";
-    case 3:
-        return "STAGGERED";
-    case 4:
-        return "POINT_INDEX";
-    case 5:
-        return "POINT_DATA";
-    case 6:
-        return "TOPOLOGY";
-    case 7:
-        return "VOXEL_VOLUME";
-    case 8:
-        return "INDEX_GRID";
-    case 9:
-        return "TENSOR_GRID";
-    default:
-        return "INVALID";
-    }
-}
-
-// Helper function to convert grid name from two uint32_t values to string
-static std::string getGridNameString(uint32_t name0, uint32_t name1)
-{
-    char nameStr[9] = { 0 }; // 8 chars + null terminator
-
-    // Pack the two uint32_t values into a char array (little-endian)
-    nameStr[0] = (char)(name0 & 0xFF);
-    nameStr[1] = (char)((name0 >> 8) & 0xFF);
-    nameStr[2] = (char)((name0 >> 16) & 0xFF);
-    nameStr[3] = (char)((name0 >> 24) & 0xFF);
-    nameStr[4] = (char)(name1 & 0xFF);
-    nameStr[5] = (char)((name1 >> 8) & 0xFF);
-    nameStr[6] = (char)((name1 >> 16) & 0xFF);
-    nameStr[7] = (char)((name1 >> 24) & 0xFF);
-
-    // Find the actual end of the string (null terminator)
-    for (int i = 0; i < 8; i++)
-    {
-        if (nameStr[i] == '\0')
-        {
-            break;
-        }
-        // Replace non-printable characters with '?'
-        if (nameStr[i] < 32 || nameStr[i] > 126)
-        {
-            nameStr[i] = '?';
-        }
-    }
-
-    return std::string(nameStr);
-}
-
-static std::string getFullGridName(pnanovdb_buf_t buf, pnanovdb_grid_handle_t grid, uint32_t flags)
-{
-    if (flags & (1 << 0)) // PNANOVDB_GRID_FLAGS_HAS_LONG_GRID_NAME
-    {
-        uint32_t blindMetadataCount = pnanovdb_grid_get_blind_metadata_count(buf, grid);
-        int64_t blindMetadataOffset = pnanovdb_grid_get_blind_metadata_offset(buf, grid);
-
-        if (blindMetadataCount > 0 && blindMetadataOffset != 0)
-        {
-            // Search through blind metadata for grid name
-            for (uint32_t i = 0; i < blindMetadataCount; ++i)
-            {
-                pnanovdb_gridblindmetadata_handle_t metadata;
-                metadata.address = pnanovdb_address_offset(
-                    grid.address, blindMetadataOffset + i * 288); // 288 = PNANOVDB_GRIDBLINDMETADATA_SIZE
-
-                uint32_t dataClass = pnanovdb_gridblindmetadata_get_data_class(buf, metadata);
-                if (dataClass == 3) // PNANOVDB_GRIDBLINDMETADATA_CLASS_GRID_NAME
-                {
-                    int64_t dataOffset = pnanovdb_gridblindmetadata_get_data_offset(buf, metadata);
-                    pnanovdb_address_t nameAddress = pnanovdb_address_offset(metadata.address, dataOffset);
-
-                    // Read the string from blind data
-                    uint64_t valueCount = pnanovdb_gridblindmetadata_get_value_count(buf, metadata);
-                    std::string longName;
-                    longName.reserve(valueCount);
-
-                    for (uint64_t j = 0; j < valueCount && j < 1024; ++j) // Safety limit
-                    {
-                        uint8_t c = pnanovdb_read_uint8(buf, pnanovdb_address_offset(nameAddress, j));
-                        if (c == '\0')
-                            break;
-                        if (c >= 32 && c <= 126) // Printable characters only
-                        {
-                            longName += (char)c;
-                        }
-                        else
-                        {
-                            longName += '?';
-                        }
-                    }
-
-                    return longName;
-                }
-            }
-        }
-
-        // Fallback: if we couldn't find the long name in blind data, show error
-        return "[Long name not found in blind data]";
-    }
-    else
-    {
-        // Standard short name (up to 256 characters)
-        uint32_t name0 = pnanovdb_grid_get_grid_name(buf, grid, 0);
-        uint32_t name1 = pnanovdb_grid_get_grid_name(buf, grid, 1);
-        return getGridNameString(name0, name1);
-    }
-}
-
-static const char* getMagicTypeName(uint64_t magic)
-{
-    if (magic == 0x304244566f6e614eULL)
-        return "NanoVDB0";
-    if (magic == 0x314244566f6e614eULL)
-        return "NanoVDB1 (Grid)";
-    if (magic == 0x324244566f6e614eULL)
-        return "NanoVDB2 (File)";
-    if ((magic & 0xFFFFFFFF) == 0x56444220UL)
-        return "OpenVDB";
-    return "Unknown";
-}
-
-static std::string getVersionString(uint32_t version)
-{
-    uint32_t major = (version >> 16) & 0xFFFF;
-    uint32_t minor = (version >> 8) & 0xFF;
-    uint32_t patch = version & 0xFF;
-
-    char versionStr[32];
-    snprintf(versionStr, sizeof(versionStr), "%u.%u.%u", major, minor, patch);
-    return std::string(versionStr);
-}
-
 static void initializeDocking()
 {
     ImGuiID dockspace_id = 1u;
@@ -365,12 +125,22 @@ static void initializeDocking()
         ImGui::DockBuilderAddNode(dockspace_id, ImGuiDockNodeFlags_DockSpace);
 
         // create dock spaces for various windows, the order matters!
+        ImGuiID dock_id_right_far = ImGui::DockBuilderSplitNode(dockspace_id, ImGuiDir_Right, 0.f, nullptr, &dockspace_id);
+        ImGui::DockBuilderSetNodeSize(dock_id_right_far, ImVec2(right_dock_width, window_height));
+        ImGui::DockBuilderDockWindow(CODE_EDITOR, dock_id_right_far);
+        ImGui::DockBuilderDockWindow(PROFILER, dock_id_right_far);
+        ImGui::DockBuilderDockWindow(FILE_HEADER, dock_id_right_far);
+
         ImGuiID dock_id_right = ImGui::DockBuilderSplitNode(dockspace_id, ImGuiDir_Right, 0.f, nullptr, &dockspace_id);
         ImGui::DockBuilderSetNodeSize(dock_id_right, ImVec2(right_dock_width, window_height));
-        ImGui::DockBuilderDockWindow(CODE_EDITOR, dock_id_right);
-        ImGui::DockBuilderDockWindow(PROFILER, dock_id_right);
-        ImGui::DockBuilderDockWindow(FILE_HEADER, dock_id_right);
-        ImGui::DockBuilderDockWindow(SCENE, dock_id_right);
+
+        ImGuiID dock_id_right_top = dock_id_right;
+        ImGuiID dock_id_right_bottom =
+            ImGui::DockBuilderSplitNode(dock_id_right_top, ImGuiDir_Down, 0.6f, nullptr, &dock_id_right_top);
+
+        ImGui::DockBuilderDockWindow(SCENE, dock_id_right_top);
+        ImGui::DockBuilderDockWindow(PROPERTIES, dock_id_right_bottom);
+        ImGui::DockBuilderDockWindow(CAMERA_VIEW, dock_id_right_bottom);
 
         ImGuiID dock_id_bottom = ImGui::DockBuilderSplitNode(dockspace_id, ImGuiDir_Down, 0.f, nullptr, &dockspace_id);
         ImGui::DockBuilderSetNodeSize(dock_id_bottom, ImVec2(window_width, bottom_dock_height));
@@ -410,6 +180,8 @@ static void createMenu(Instance* ptr)
             ImGui::MenuItem(CONSOLE, "", &ptr->window.show_console);
             ImGui::MenuItem(SHADER_PARAMS, "", &ptr->window.show_shader_params);
             ImGui::MenuItem(SCENE, "", &ptr->window.show_scene);
+            ImGui::MenuItem(CAMERA_VIEW, "", &ptr->window.show_camera_view);
+            ImGui::MenuItem(PROPERTIES, "", &ptr->window.show_scene_properties);
             ImGui::MenuItem(BENCHMARK, "", &ptr->window.show_benchmark);
             ImGui::EndMenu();
         }
@@ -426,1411 +198,22 @@ static void createMenu(Instance* ptr)
 
 static void showWindows(Instance* ptr, float delta_time)
 {
-    if (ptr->window.show_viewport_settings)
-    {
-        if (ImGui::Begin(VIEWPORT_SETTINGS, &ptr->window.show_viewport_settings))
-        {
-            // viewport options
-            {
-                ImGui::BeginGroup();
-                ViewportOption selectedOption = ptr->viewport_option;
-                if (ImGui::RadioButton("NanoVDB", selectedOption == ViewportOption::NanoVDB))
-                {
-                    ptr->viewport_option = ViewportOption::NanoVDB;
-                    ptr->render_settings_name = ptr->viewport_settings[(int)ptr->viewport_option].render_settings_name;
-                    ptr->pending.load_camera = true;
-                }
-                ImGui::SameLine();
-                if (ImGui::RadioButton("Raster2D", selectedOption == ViewportOption::Raster2D))
-                {
-                    ptr->viewport_option = ViewportOption::Raster2D;
-                    ptr->render_settings_name = ptr->viewport_settings[(int)ptr->viewport_option].render_settings_name;
-                    ptr->pending.load_camera = true;
-                }
-                ImGui::EndGroup();
-            }
-            {
-                ImGui::BeginGroup();
-                if (ImGui::BeginCombo("Viewport Camera", "Select..."))
-                {
-                    for (const auto& pair : ptr->saved_render_settings)
-                    {
-                        bool is_selected = (ptr->render_settings_name == pair.first);
-                        if (ImGui::Selectable(pair.first.c_str(), is_selected))
-                        {
-                            ptr->render_settings_name = pair.first;
-                            ptr->viewport_settings[(int)ptr->viewport_option].render_settings_name =
-                                ptr->render_settings_name;
-                            ptr->pending.load_camera = true;
+    showViewportSettingsWindow(ptr);
+    showRenderSettingsWindow(ptr);
+    showCompilerSettingsWindow(ptr);
 
-                            // clear selected debug camera when switching to saved camera
-                            ptr->selected_camera_view.clear();
+    showShaderParamsWindow(ptr);
+    showBenchmarkWindow(ptr);
 
-                            ImGui::MarkIniSettingsDirty();
-                        }
-                        if (is_selected)
-                        {
-                            ImGui::SetItemDefaultFocus();
-                        }
-                    }
-                    ImGui::EndCombo();
-                }
-                ImGui::InputText("##render_settings_name", &ptr->render_settings_name);
-                ImGui::SameLine();
-                if (ImGui::Button("Save"))
-                {
-                    if (!ptr->render_settings_name.empty())
-                    {
-                        if (ptr->saved_render_settings.find(ptr->render_settings_name) == ptr->saved_render_settings.end())
-                        {
-                            // save camera state in editor update loop
-                            ptr->pending.save_camera = true;
-                        }
-                        else
-                        {
-                            pnanovdb_editor::Console::getInstance().addLog(
-                                "Render settings '%s' already exists", ptr->render_settings_name.c_str());
-                        }
-                    }
-                    else
-                    {
-                        pnanovdb_editor::Console::getInstance().addLog("Please enter a name for the render settings");
-                    }
-                }
-                ImGui::SameLine();
-                if (ImGui::Button("Remove"))
-                {
-                    if (ptr->saved_render_settings.erase(ptr->render_settings_name))
-                    {
-                        pnanovdb_editor::Console::getInstance().addLog(
-                            "Render settings '%s' removed", ptr->render_settings_name.c_str());
-                        ptr->render_settings_name = "";
-                    }
-                    else
-                    {
-                        pnanovdb_editor::Console::getInstance().addLog(
-                            "Render settings '%s' not found", ptr->render_settings_name.c_str());
-                    }
-                }
-                ImGui::EndGroup();
-            }
-            ImGui::Separator();
+    showSceneWindow(ptr);
+    showCameraViewWindow(ptr);
+    showPropertiesWindow(ptr);
 
-            ImGui::Text("Gaussian File");
-            {
-                ImGui::BeginGroup();
-                ImGui::InputText("##viewport_raster_file", &ptr->raster_filepath);
-                ImGui::SameLine();
-                if (ImGui::Button("...##open_raster_file"))
-                {
-                    ptr->pending.find_raster_file = true;
+    showFileHeaderWindow(ptr);
+    showCodeEditorWindow(ptr);
+    showProfilerWindow(ptr, delta_time);
 
-                    IGFD::FileDialogConfig config;
-                    config.path = ".";
-
-                    ImGuiFileDialog::Instance()->OpenDialog("OpenRasterFileDlgKey", "Open Gaussian File",
-                                                            "Gaussian Files (*.npy *.npz *.ply){.npy,.npz,.ply}", config);
-                }
-                ImGui::SameLine();
-                if (ImGui::Button("Show##Gaussian"))
-                {
-                    // runs rasterization on a worker thread first
-                    ptr->pending.update_raster = true;
-                }
-                if (ptr->viewport_option == ViewportOption::NanoVDB)
-                {
-                    ImGui::InputFloat("VoxelsPerUnit", &ptr->raster_voxels_per_unit);
-                }
-                ImGui::EndGroup();
-            }
-
-            if (ptr->viewport_option == ViewportOption::NanoVDB)
-            {
-                ImGui::Text("NanoVDB File");
-                {
-                    ImGui::BeginGroup();
-                    ImGui::InputText("##viewport_nanovdb_file", &ptr->nanovdb_filepath);
-                    ImGui::SameLine();
-                    if (ImGui::Button("...##open_nanovddb_file"))
-                    {
-                        ptr->pending.open_file = true;
-                    }
-                    ImGui::SameLine();
-                    if (ImGui::Button("Show##NanoVDB"))
-                    {
-                        // TODO: does it need to be on a worker thread?
-                        pnanovdb_editor::Console::getInstance().addLog(
-                            "Opening file '%s'", ptr->nanovdb_filepath.c_str());
-                        ptr->pending.load_nvdb = true;
-                    }
-                    ImGui::EndGroup();
-                }
-            }
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_render_settings)
-    {
-        if (ImGui::Begin(RENDER_SETTINGS, &ptr->window.show_render_settings))
-        {
-            ImGui::Text("Viewport Shader");
-            {
-                ImGui::BeginGroup();
-                if (ImGui::BeginCombo("##viewport_shader", "Select..."))
-                {
-                    for (const auto& shader : ptr->viewport_shaders)
-                    {
-                        if (ImGui::Selectable(shader.c_str()))
-                        {
-                            ptr->pending.viewport_shader_name = shader;
-                        }
-                    }
-                    ImGui::EndCombo();
-                }
-                ImGui::InputText("##viewport_shader_name", &ptr->pending.viewport_shader_name);
-                ImGui::SameLine();
-                if (ImGui::Button("Update"))
-                {
-                    pnanovdb_editor::CodeEditor::getInstance().setSelectedShader(ptr->pending.viewport_shader_name);
-                    ptr->shader_name = ptr->pending.viewport_shader_name;
-                    ptr->pending.update_shader = true;
-                }
-                ImGui::EndGroup();
-            }
-            ImGui::Separator();
-
-            auto settings = ptr->render_settings;
-
-            IMGUI_CHECKBOX_SYNC("VSync", settings->vsync);
-            IMGUI_CHECKBOX_SYNC("Projection RH", settings->is_projection_rh);
-            IMGUI_CHECKBOX_SYNC("Orthographic", settings->is_orthographic);
-            IMGUI_CHECKBOX_SYNC("Reverse Z", settings->is_reverse_z);
-            {
-                int up_axis = settings->is_y_up ? 0 : 1;
-                const char* up_axis_items[] = { "Y", "Z" };
-                if (ImGui::Combo("Up Axis", &up_axis, up_axis_items, IM_ARRAYSIZE(up_axis_items)))
-                {
-                    settings->is_y_up = (up_axis == 0);
-                }
-
-                IMGUI_CHECKBOX_SYNC("Upside down", settings->is_upside_down);
-            }
-            ImGui::DragFloat("Shift Speed Multiplier", &settings->key_translation_shift_multiplier, 0.f, 1.f, 10000.f,
-                             "%.1f", ImGuiSliderFlags_Logarithmic | ImGuiSliderFlags_AlwaysClamp);
-            ImGui::Separator();
-
-            IMGUI_CHECKBOX_SYNC("Video Encode", settings->enable_encoder);
-            if (ImGui::BeginCombo("Resolution", "Select..."))
-            {
-                const char* labels[4] = { "1440x720", "1920x1080", "2560x1440", "3840x2160" };
-                const int widths[4] = { 1440, 1920, 2560, 3840 };
-                const int heights[4] = { 720, 1080, 1440, 2160 };
-                for (int i = 0; i < 4; i++)
-                {
-                    if (ImGui::Selectable(labels[i]))
-                    {
-                        settings->encode_width = widths[i];
-                        settings->encode_height = heights[i];
-                    }
-                }
-                ImGui::EndCombo();
-            }
-            ImGui::InputText("Server Address", settings->server_address, 256u);
-            ImGui::InputInt("Server Port", &settings->server_port);
-
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_compiler_settings)
-    {
-        if (ImGui::Begin(COMPILER_SETTINGS, &ptr->window.show_compiler_settings))
-        {
-            std::lock_guard<std::mutex> lock(ptr->compiler_settings_mutex);
-
-            IMGUI_CHECKBOX_SYNC("Row Major Matrix Layout", ptr->compiler_settings.is_row_major);
-            IMGUI_CHECKBOX_SYNC("Create HLSL Output", ptr->compiler_settings.hlsl_output);
-            // TODO: add downstream compilers dropdown
-            IMGUI_CHECKBOX_SYNC("Use glslang", ptr->compiler_settings.use_glslang);
-
-            const char* compile_targets[] = { "Select...", "Vulkan", "CPU" };
-            int compile_target = (int)ptr->compiler_settings.compile_target;
-            if (ImGui::Combo("Target", &compile_target, compile_targets, IM_ARRAYSIZE(compile_targets)))
-            {
-                ptr->compiler_settings.compile_target = (pnanovdb_compile_target_type_t)(compile_target);
-                // Clear entry point name, default is assigned in compiler
-                ptr->compiler_settings.entry_point_name[0] = '\0';
-            }
-            ImGui::InputText("Entry Point Name", ptr->compiler_settings.entry_point_name,
-                             IM_ARRAYSIZE(ptr->compiler_settings.entry_point_name));
-            if (ImGui::IsItemHovered())
-            {
-                ImGui::SetTooltip("Empty uses default");
-            }
-
-            ImGui::Separator();
-            ImGui::Text("Additional Shader Directories");
-            if (!ptr->additional_shader_directories.empty())
-            {
-                for (size_t i = 0; i < ptr->additional_shader_directories.size(); i++)
-                {
-                    ImGui::PushID(static_cast<int>(i));
-                    ImGui::Text("  %s", ptr->additional_shader_directories[i].c_str());
-                    ImGui::SameLine();
-                    if (ImGui::Button("-"))
-                    {
-                        std::string dirToRemove = ptr->additional_shader_directories[i];
-                        pnanovdb_editor::ShaderMonitor::getInstance().removePath(dirToRemove);
-                        ptr->additional_shader_directories.erase(ptr->additional_shader_directories.begin() + i);
-                        pnanovdb_editor::Console::getInstance().addLog(
-                            "Removed shader directory: %s", dirToRemove.c_str());
-                        ImGui::MarkIniSettingsDirty();
-                        i--;
-                    }
-                    ImGui::PopID();
-                }
-            }
-
-            ImGui::InputText("##shader_directory", &ptr->pending_shader_directory);
-            ImGui::SameLine();
-            if (ImGui::Button("...##select_shader_directory"))
-            {
-                ptr->pending.find_shader_directory = true;
-            }
-            ImGui::SameLine();
-            if (ImGui::Button("+"))
-            {
-                if (!ptr->pending_shader_directory.empty())
-                {
-                    std::filesystem::path dirPath(ptr->pending_shader_directory);
-                    if (std::filesystem::exists(dirPath) && std::filesystem::is_directory(dirPath))
-                    {
-                        auto it = std::find(ptr->additional_shader_directories.begin(),
-                                            ptr->additional_shader_directories.end(), ptr->pending_shader_directory);
-                        if (it == ptr->additional_shader_directories.end())
-                        {
-                            ptr->additional_shader_directories.push_back(ptr->pending_shader_directory);
-
-                            auto callback = pnanovdb_editor::get_shader_recompile_callback(ptr, ptr->compiler);
-
-                            pnanovdb_editor::monitor_shader_dir(ptr->pending_shader_directory.c_str(), callback);
-                            pnanovdb_editor::Console::getInstance().addLog(
-                                "Added shader directory: %s", ptr->pending_shader_directory.c_str());
-                            ImGui::MarkIniSettingsDirty();
-                        }
-                        else
-                        {
-                            pnanovdb_editor::Console::getInstance().addLog(
-                                "Directory already being monitored: %s", ptr->pending_shader_directory.c_str());
-                        }
-                        ptr->pending_shader_directory.clear();
-                    }
-                    else
-                    {
-                        pnanovdb_editor::Console::getInstance().addLog(
-                            "Invalid directory path: %s", ptr->pending_shader_directory.c_str());
-                    }
-                }
-            }
-
-            const std::string compiledShadersDir = pnanovdb_shader::getShaderCacheDir();
-            if (!compiledShadersDir.empty())
-            {
-                ImGui::Text("Shader Cache");
-                ImGui::Text("  %s", compiledShadersDir.c_str());
-                if (ImGui::Button("Clear"))
-                {
-                    std::filesystem::path shaderDir(compiledShadersDir);
-                    if (std::filesystem::exists(shaderDir) && std::filesystem::is_directory(shaderDir))
-                    {
-                        for (const auto& entry : std::filesystem::directory_iterator(shaderDir))
-                        {
-                            std::filesystem::remove_all(entry.path());
-                        }
-                        pnanovdb_editor::Console::getInstance().addLog("Shader cache cleared");
-                    }
-                }
-            }
-
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_shader_params)
-    {
-        if (ImGui::Begin(SHADER_PARAMS, &ptr->window.show_shader_params))
-        {
-            std::string shader_name;
-
-            ImGui::BeginGroup();
-
-            // Show params which are parsed from the shader used in current vieweport
-            if (ImGui::RadioButton(
-                    "Viewport Shader", ptr->pending.shader_selection_mode == ShaderSelectionMode::UseViewportShader))
-            {
-                ptr->pending.shader_selection_mode = ShaderSelectionMode::UseViewportShader;
-            }
-
-            // Show params which are parsed from the shader opened in the code editor
-            if (ImGui::RadioButton("Shader Editor Selected:",
-                                   ptr->pending.shader_selection_mode == ShaderSelectionMode::UseCodeEditorShader))
-            {
-                ptr->pending.shader_selection_mode = ShaderSelectionMode::UseCodeEditorShader;
-            }
-            ImGui::SameLine();
-            shader_name = pnanovdb_editor::CodeEditor::getInstance().getSelectedShader().c_str();
-            // strip the directory, show only the filename
-            std::filesystem::path fsPath(shader_name);
-            ImGui::Text("%s", fsPath.filename().string().c_str());
-
-            // Show params for a specific shader group
-            // TODO: This will belong to the Pipeline window
-            bool is_group_mode = ptr->pending.shader_selection_mode == ShaderSelectionMode::UseShaderGroup;
-            if (ImGui::RadioButton("Shader Group", is_group_mode))
-            {
-                ptr->pending.shader_selection_mode = ShaderSelectionMode::UseShaderGroup;
-            }
-            ImGui::SameLine();
-            if (ImGui::InputText("##shader_group", &ptr->shader_group))
-            {
-                ImGui::MarkIniSettingsDirty();
-            }
-            ImGui::EndGroup();
-
-            const std::string& target_name = is_group_mode ? ptr->shader_group : shader_name;
-            if (!target_name.empty())
-            {
-                if (ptr->shader_params.isJsonLoaded(target_name, is_group_mode))
-                {
-                    if (ImGui::Button("Reload JSON"))
-                    {
-                        if (is_group_mode)
-                        {
-                            if (ptr->shader_params.loadGroup(target_name, true))
-                            {
-                                pnanovdb_editor::Console::getInstance().addLog(
-                                    "Group params '%s' updated", target_name.c_str());
-                            }
-                            else
-                            {
-                                pnanovdb_editor::Console::getInstance().addLog(
-                                    "Failed to reload group params '%s'", target_name.c_str());
-                            }
-                        }
-                        else
-                        {
-                            pnanovdb_editor::CodeEditor::getInstance().saveShaderParams();
-                            if (ptr->shader_params.load(target_name, true))
-                            {
-                                pnanovdb_editor::Console::getInstance().addLog(
-                                    "Shader params for '%s' updated", target_name.c_str());
-                            }
-                            else
-                            {
-                                pnanovdb_editor::Console::getInstance().addLog(
-                                    "Failed to reload shader params for '%s'", target_name.c_str());
-                            }
-                        }
-                    }
-                }
-                else
-                {
-                    if (ImGui::Button("Create JSON"))
-                    {
-                        if (ptr->pending.shader_selection_mode == ShaderSelectionMode::UseShaderGroup)
-                        {
-                            ptr->shader_params.createGroup(target_name);
-                        }
-                        else
-                        {
-                            ptr->shader_params.create(target_name);
-                            pnanovdb_editor::CodeEditor::getInstance().updateViewer();
-                        }
-                    }
-                }
-            }
-
-            ImGui::Separator();
-
-            if (!target_name.empty())
-            {
-                if (ptr->pending.shader_selection_mode == ShaderSelectionMode::UseShaderGroup)
-                {
-                    ptr->shader_params.renderGroup(target_name);
-                }
-                else
-                {
-                    if (ptr->shader_params.isJsonLoaded(target_name))
-                    {
-                        ptr->shader_params.render(target_name);
-                    }
-                    else
-                    {
-                        ImGui::PushStyleColor(ImGuiCol_Text, IM_COL32(191, 97, 106, 255)); // Red color
-                        ImGui::TextWrapped("Shader params JSON is missing");
-                        ImGui::PopStyleColor();
-                    }
-                }
-            }
-            else
-            {
-                ImGui::TextDisabled("No shader selected");
-            }
-
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_scene)
-    {
-        if (ImGui::Begin(SCENE, &ptr->window.show_scene))
-        {
-            ImVec2 fullAvail = ImGui::GetContentRegionAvail();
-            static float topHeight = -1.f; // will be initialized on first frame
-            const float minTop = 100.f;
-            const float minBottom = 80.f;
-            if (topHeight < 0.f && fullAvail.y > minTop + minBottom)
-            {
-                topHeight = fullAvail.y * 0.333f;
-            }
-            else if (fullAvail.y > 0)
-            {
-                topHeight = std::max(minTop, std::min(topHeight, fullAvail.y - minTop));
-            }
-
-            // Top: Tree view
-            if (ImGui::BeginChild("##SceneTop", ImVec2(0, topHeight), true))
-            {
-                if (ImGui::TreeNodeEx("Cameras", ImGuiTreeNodeFlags_DefaultOpen))
-                {
-                    if (!ptr->camera_views)
-                    {
-                        ImGui::TextDisabled("No camera views available");
-                    }
-                    else if (ptr->camera_views->empty())
-                    {
-                        ImGui::TextDisabled("No camera views added");
-                    }
-                    else
-                    {
-                        for (auto& cameraPair : *ptr->camera_views)
-                        {
-                            pnanovdb_camera_view_t* camera = cameraPair.second;
-                            if (!camera)
-                            {
-                                continue;
-                            }
-                            const std::string& cameraName = cameraPair.first;
-                            IMGUI_CHECKBOX_SYNC(("##Visible" + cameraName).c_str(), camera->is_visible);
-                            ImGui::SameLine();
-                            ImGuiTreeNodeFlags flags = ImGuiTreeNodeFlags_Leaf | ImGuiTreeNodeFlags_NoTreePushOnOpen;
-                            bool isSelected = (ptr->selected_camera_frustum == cameraName);
-                            if (isSelected)
-                            {
-                                flags |= ImGuiTreeNodeFlags_Selected;
-                            }
-                            ImGui::TreeNodeEx(cameraName.c_str(), flags);
-                            if (ImGui::IsItemClicked())
-                            {
-                                ptr->selected_camera_frustum = cameraName;
-                                ptr->selected_view_type = ViewsTypes::Cameras;
-                            }
-                        }
-                    }
-                    ImGui::TreePop();
-                }
-
-                if (ImGui::TreeNodeEx("Gaussian Scenes", ImGuiTreeNodeFlags_DefaultOpen))
-                {
-                    if (!ptr->gaussian_views)
-                    {
-                        ImGui::TextDisabled("No gaussian scenes available");
-                    }
-                    else if (ptr->gaussian_views->empty())
-                    {
-                        ImGui::TextDisabled("No gaussian scenes added");
-                    }
-                    else
-                    {
-                        for (auto& gaussianPair : *ptr->gaussian_views)
-                        {
-                            const std::string& gaussianName = gaussianPair.first;
-                            bool isSelected = (ptr->selected_gaussian_view == gaussianName);
-
-                            if (ImGui::RadioButton(("##Radio" + gaussianName).c_str(), isSelected))
-                            {
-                                ptr->pending.viewport_gaussian_view = gaussianName;
-                                ptr->selected_view_type = ViewsTypes::GaussianScenes;
-                            }
-                            ImGui::SameLine();
-
-                            ImGuiTreeNodeFlags flags = ImGuiTreeNodeFlags_Leaf | ImGuiTreeNodeFlags_NoTreePushOnOpen;
-                            if (isSelected)
-                            {
-                                flags |= ImGuiTreeNodeFlags_Selected;
-                            }
-                            ImGui::TreeNodeEx(gaussianName.c_str(), flags);
-                            if (ImGui::IsItemClicked())
-                            {
-                                ptr->pending.viewport_gaussian_view = gaussianName;
-                                ptr->selected_view_type = ViewsTypes::GaussianScenes;
-                            }
-                        }
-                    }
-                    ImGui::TreePop();
-                }
-            }
-            ImGui::EndChild();
-
-            // Splitter
-            ImRect bb;
-            bb.Min = ImGui::GetWindowPos() + ImGui::GetCursorPos();
-            bb.Min.y += 2.f;
-            bb.Max = bb.Min + ImVec2(fullAvail.x, 2.f);
-            float bottomHeight = fullAvail.y - topHeight;
-            ImGui::SplitterBehavior(
-                bb, ImGui::GetID("##SceneSplitter"), ImGuiAxis_Y, &topHeight, &bottomHeight, minTop, minBottom);
-            ImGui::Spacing();
-
-            // Bottom: Property tab
-            if (ImGui::BeginChild("##SceneBottom", ImVec2(0, 0), true))
-            {
-                if (ptr->selected_view_type == ViewsTypes::Cameras)
-                {
-                    if (ptr->selected_camera_frustum.empty())
-                    {
-                        ImGui::TextDisabled("No camera selected");
-                    }
-                    else
-                    {
-                        auto it = ptr->camera_views ? ptr->camera_views->find(ptr->selected_camera_frustum) :
-                                                      std::map<std::string, pnanovdb_camera_view_t*>().end();
-                        if (it != (ptr->camera_views ? ptr->camera_views->end() :
-                                                       std::map<std::string, pnanovdb_camera_view_t*>().end()))
-                        {
-                            pnanovdb_camera_view_t* camera = it->second;
-                            if (camera)
-                            {
-                                ImGui::Text("Total Cameras: %d", camera->num_cameras);
-                                int maxIndex = (camera->num_cameras > 0) ? ((int)camera->num_cameras - 1) : 0;
-                                if (maxIndex > 0)
-                                {
-                                    ImGui::SliderInt("Camera Index",
-                                                     &ptr->camera_frustum_index[ptr->selected_camera_frustum], 0,
-                                                     maxIndex, "%d");
-                                }
-                                else
-                                {
-                                    ImGui::Text("Camera Index: 0");
-                                    ImGui::Dummy(ImVec2(0.f, 1.f));
-                                }
-                                int cameraIdx = ptr->camera_frustum_index[ptr->selected_camera_frustum];
-                                if (ImGui::Button("Set Viewport Camera"))
-                                {
-                                    pnanovdb_vec3_t& up = camera->states[cameraIdx].eye_up;
-                                    if (camera->states[cameraIdx].eye_up.x != ptr->render_settings->camera_state.eye_up.x ||
-                                        camera->states[cameraIdx].eye_up.y != ptr->render_settings->camera_state.eye_up.y ||
-                                        camera->states[cameraIdx].eye_up.z != ptr->render_settings->camera_state.eye_up.z)
-                                    {
-                                        pnanovdb_vec3_t& dir = camera->states[cameraIdx].eye_direction;
-                                        pnanovdb_vec3_t right = { dir.y * up.z - dir.z * up.y,
-                                                                  dir.z * up.x - dir.x * up.z,
-                                                                  dir.x * up.y - dir.y * up.x };
-                                        up.x = -(right.y * dir.z - right.z * dir.y);
-                                        up.y = -(right.z * dir.x - right.x * dir.z);
-                                        up.z = -(right.x * dir.y - right.y * dir.x);
-                                        float len = sqrtf(up.x * up.x + up.y * up.y + up.z * up.z);
-                                        if (len > EPSILON)
-                                        {
-                                            up.x /= len;
-                                            up.y /= len;
-                                            up.z /= len;
-                                        }
-                                    }
-                                    ptr->render_settings->camera_state = camera->states[cameraIdx];
-                                    ptr->render_settings->camera_state.eye_up = up;
-                                    ptr->render_settings->camera_config = camera->configs[cameraIdx];
-                                    ptr->render_settings->is_projection_rh = camera->configs[cameraIdx].is_projection_rh;
-                                    ptr->render_settings->is_orthographic = camera->configs[cameraIdx].is_orthographic;
-                                    ptr->render_settings->is_reverse_z = camera->configs[cameraIdx].is_reverse_z;
-                                    ptr->render_settings->sync_camera = PNANOVDB_TRUE;
-                                }
-
-                                pnanovdb_vec3_t eyePosition =
-                                    pnanovdb_camera_get_eye_position_from_state(&camera->states[cameraIdx]);
-                                float eyePos[3] = { eyePosition.x, eyePosition.y, eyePosition.z };
-                                if (ImGui::DragFloat3("Origin", eyePos, 0.1f))
-                                {
-                                    pnanovdb_camera_state_t* state = &camera->states[cameraIdx];
-                                    state->position.x =
-                                        eyePos[0] + state->eye_direction.x * state->eye_distance_from_position;
-                                    state->position.y =
-                                        eyePos[1] + state->eye_direction.y * state->eye_distance_from_position;
-                                    state->position.z =
-                                        eyePos[2] + state->eye_direction.z * state->eye_distance_from_position;
-                                }
-
-                                pnanovdb_camera_state_t* state = &camera->states[cameraIdx];
-                                float lookAt[3] = { state->position.x, state->position.y, state->position.z };
-                                if (ImGui::DragFloat3("Look At", lookAt, 0.1f))
-                                {
-                                    pnanovdb_vec3_t eye = pnanovdb_camera_get_eye_position_from_state(state);
-                                    pnanovdb_vec3_t delta = { lookAt[0] - eye.x, lookAt[1] - eye.y, lookAt[2] - eye.z };
-                                    float len2 = delta.x * delta.x + delta.y * delta.y + delta.z * delta.z;
-                                    if (len2 > 0.f)
-                                    {
-                                        float len = pnanovdb_camera_sqrt(len2);
-                                        pnanovdb_vec3_t dir = { delta.x / len, delta.y / len, delta.z / len };
-                                        state->position.x = lookAt[0];
-                                        state->position.y = lookAt[1];
-                                        state->position.z = lookAt[2];
-                                        state->eye_direction = dir;
-                                        state->eye_distance_from_position = len;
-                                    }
-                                    else
-                                    {
-                                        state->position.x = lookAt[0];
-                                        state->position.y = lookAt[1];
-                                        state->position.z = lookAt[2];
-                                        state->eye_distance_from_position = 0.f;
-                                    }
-                                }
-                                float upVec[3] = { state->eye_up.x, state->eye_up.y, state->eye_up.z };
-                                if (ImGui::DragFloat3("Up", upVec, 0.1f))
-                                {
-                                    state->eye_up.x = upVec[0];
-                                    state->eye_up.y = upVec[1];
-                                    state->eye_up.z = upVec[2];
-                                    float len =
-                                        sqrtf(state->eye_up.x * state->eye_up.x + state->eye_up.y * state->eye_up.y +
-                                              state->eye_up.z * state->eye_up.z);
-                                    if (len > EPSILON)
-                                    {
-                                        state->eye_up.x /= len;
-                                        state->eye_up.y /= len;
-                                        state->eye_up.z /= len;
-                                    }
-                                }
-                                ImGui::Separator();
-                                ImGui::DragFloat("Axis Length", &camera->axis_length, 1.f, 0.f, 100.f);
-                                ImGui::DragFloat("Axis Thickness", &camera->axis_thickness, 0.1f, 0.f, 10.f);
-                                ImGui::DragFloat("Frustum Line Width", &camera->frustum_line_width, 0.1f, 0.f, 10.f);
-                                ImGui::DragFloat("Frustum Scale", &camera->frustum_scale, 0.1f, 0.f, 10.f);
-                                float frustumColor[4] = { (float)camera->frustum_color.x, (float)camera->frustum_color.y,
-                                                          (float)camera->frustum_color.z, 1.0f };
-                                if (ImGui::ColorEdit4("Frustum Color", frustumColor))
-                                {
-                                    camera->frustum_color.x = frustumColor[0];
-                                    camera->frustum_color.y = frustumColor[1];
-                                    camera->frustum_color.z = frustumColor[2];
-                                }
-                                ImGui::Separator();
-                                ImGui::DragFloat(
-                                    "Near Plane", &camera->configs[cameraIdx].near_plane, 0.1f, 0.01f, 10000.f);
-                                ImGui::DragFloat("Far Plane", &camera->configs[cameraIdx].far_plane, 10.f, 1.f, 100000.f);
-                                if (camera->configs[cameraIdx].is_orthographic)
-                                {
-                                    ImGui::DragFloat("Orthographic Y", &camera->configs[cameraIdx].orthographic_y, 0.1f,
-                                                     0.f, 100000.f);
-                                }
-                                else
-                                {
-                                    ImGui::DragFloat("FOV", &camera->configs[cameraIdx].fov_angle_y, 0.01f, 0.f, M_PI_2);
-                                    ImGui::DragFloat(
-                                        "Aspect Ratio", &camera->configs[cameraIdx].aspect_ratio, 0.01f, 0.f, 2.f);
-                                }
-                            }
-                        }
-                    }
-                }
-                else if (ptr->selected_view_type == ViewsTypes::GaussianScenes)
-                {
-                    if (ptr->selected_gaussian_view.empty())
-                    {
-                        ImGui::TextDisabled("No gaussian scene selected");
-                    }
-                    else
-                    {
-                        ptr->shader_params.renderGroup(imgui_instance_user::s_raster2d_shader_group);
-                    }
-                }
-                else
-                {
-                    ImGui::TextDisabled("No item selected");
-                }
-            }
-            ImGui::EndChild();
-
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_benchmark)
-    {
-        if (ImGui::Begin(BENCHMARK, &ptr->window.show_benchmark))
-        {
-            if (ImGui::Button("Run Benchmark"))
-            {
-                // For now, guess the ref nvdb by convention
-                std::string ref_nvdb = ptr->nanovdb_filepath;
-                std::string suffix = "_node2.nvdb";
-                if (ref_nvdb.size() > suffix.size())
-                {
-                    ref_nvdb.erase(ref_nvdb.size() - suffix.size(), suffix.size());
-                }
-                ref_nvdb.append(".nvdb");
-                printf("nanovdbFile_(%s) ref_nvdb(%s) suffix(%s)\n", ptr->nanovdb_filepath.c_str(), ref_nvdb.c_str(),
-                       suffix.c_str());
-                pnanovdb_editor::node2_verify_gpu(ref_nvdb.c_str(), ptr->nanovdb_filepath.c_str());
-            }
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_file_header)
-    {
-        if (ImGui::Begin(FILE_HEADER, &ptr->window.show_file_header))
-        {
-            if (ptr->nanovdb_array && ptr->nanovdb_array->data && ptr->nanovdb_array->element_count > 0)
-            {
-                pnanovdb_buf_t buf =
-                    pnanovdb_make_buf((pnanovdb_uint32_t*)ptr->nanovdb_array->data, ptr->nanovdb_array->element_count);
-                pnanovdb_grid_handle_t grid = { pnanovdb_address_null() };
-
-                ImGui::Text("NanoVDB Grid Information");
-                ImGui::Separator();
-
-                if (ImGui::CollapsingHeader("Grid Header", ImGuiTreeNodeFlags_DefaultOpen))
-                {
-                    uint64_t magic = pnanovdb_grid_get_magic(buf, grid);
-                    ImGui::Text("Magic: %s", getMagicTypeName(magic));
-
-                    uint32_t version = pnanovdb_grid_get_version(buf, grid);
-                    std::string versionStr = getVersionString(version);
-                    ImGui::Text("Version: %s", versionStr.c_str());
-
-                    uint32_t flags = pnanovdb_grid_get_flags(buf, grid);
-                    if (ImGui::CollapsingHeader("Grid Flags"))
-                    {
-                        ImGui::Indent();
-
-                        struct FlagInfo
-                        {
-                            uint32_t bit;
-                            const char* name;
-                            const char* description;
-                        };
-
-                        FlagInfo flagInfos[] = { { 1 << 0, "Long Grid Name", "Grid name > 256 characters" },
-                                                 { 1 << 1, "Bounding Box", "Nodes contain bounding boxes" },
-                                                 { 1 << 2, "Min/Max Values", "Nodes contain min/max statistics" },
-                                                 { 1 << 3, "Average Values", "Nodes contain average statistics" },
-                                                 { 1 << 4, "Std Deviation", "Nodes contain standard deviation" },
-                                                 { 1 << 5, "Breadth First", "Memory layout is breadth-first" } };
-
-                        bool hasAnyFlags = false;
-                        for (const auto& flagInfo : flagInfos)
-                        {
-                            bool isSet = (flags & flagInfo.bit) != 0;
-                            if (isSet)
-                                hasAnyFlags = true;
-
-                            // Light green or gray
-                            ImU32 textColor = isSet ? IM_COL32(144, 238, 144, 255) : IM_COL32(128, 128, 128, 255);
-                            ImGui::PushStyleColor(ImGuiCol_Text, textColor);
-
-                            ImGui::Text("[%s] %s", isSet ? "X" : " ", flagInfo.name);
-                            if (ImGui::IsItemHovered())
-                            {
-                                ImGui::SetTooltip("%s", flagInfo.description);
-                            }
-                            ImGui::PopStyleColor();
-                        }
-
-                        if (!hasAnyFlags)
-                        {
-                            ImGui::PushStyleColor(ImGuiCol_Text, IM_COL32(128, 128, 128, 255));
-                            ImGui::Text("[ ] No flags set");
-                            ImGui::PopStyleColor();
-                        }
-
-                        ImGui::Unindent();
-                    }
-                    ImGui::Text("Grid Index: %u", pnanovdb_grid_get_grid_index(buf, grid));
-                    ImGui::Text("Grid Count: %u", pnanovdb_grid_get_grid_count(buf, grid));
-                    ImGui::Text("Grid Size: %llu bytes", (unsigned long long)pnanovdb_grid_get_grid_size(buf, grid));
-                    uint32_t gridType = pnanovdb_grid_get_grid_type(buf, grid);
-                    ImGui::Text("Grid Type: %s", getGridTypeName(gridType));
-                    uint32_t gridClass = pnanovdb_grid_get_grid_class(buf, grid);
-                    ImGui::Text("Grid Class: %s", getGridClassName(gridClass));
-
-                    std::string gridNameStr = getFullGridName(buf, grid, flags);
-                    ImGui::Text("Grid Name: \"%s\"", gridNameStr.c_str());
-                    if (flags & (1 << 0)) // HasLongGridName flag
-                    {
-                        ImGui::SameLine();
-                        ImGui::TextColored(ImVec4(0.5f, 0.7f, 1.0f, 1.0f), "(Long Name)");
-                        if (ImGui::IsItemHovered())
-                        {
-                            ImGui::SetTooltip("Grid name longer than 256 characters, stored in blind metadata");
-                        }
-                    }
-
-                    ImGui::Text("World BBox Min: (%.3f, %.3f, %.3f)", pnanovdb_grid_get_world_bbox(buf, grid, 0),
-                                pnanovdb_grid_get_world_bbox(buf, grid, 1), pnanovdb_grid_get_world_bbox(buf, grid, 2));
-                    ImGui::Text("World BBox Max: (%.3f, %.3f, %.3f)", pnanovdb_grid_get_world_bbox(buf, grid, 3),
-                                pnanovdb_grid_get_world_bbox(buf, grid, 4), pnanovdb_grid_get_world_bbox(buf, grid, 5));
-
-                    ImGui::Text("Voxel Size: (%.6f, %.6f, %.6f)", pnanovdb_grid_get_voxel_size(buf, grid, 0),
-                                pnanovdb_grid_get_voxel_size(buf, grid, 1), pnanovdb_grid_get_voxel_size(buf, grid, 2));
-                }
-
-                if (ImGui::CollapsingHeader("Tree Information", ImGuiTreeNodeFlags_DefaultOpen))
-                {
-                    pnanovdb_tree_handle_t tree = pnanovdb_grid_get_tree(buf, grid);
-                    ImGui::Text("Active Voxel Count: %llu", (unsigned long long)pnanovdb_tree_get_voxel_count(buf, tree));
-                    ImGui::Text("Leaf Node Count: %u", pnanovdb_tree_get_node_count_leaf(buf, tree));
-                    ImGui::Text("Lower Internal Node Count: %u", pnanovdb_tree_get_node_count_lower(buf, tree));
-                    ImGui::Text("Upper Internal Node Count: %u", pnanovdb_tree_get_node_count_upper(buf, tree));
-                    ImGui::Text("Lower Tile Count: %u", pnanovdb_tree_get_tile_count_lower(buf, tree));
-                    ImGui::Text("Upper Tile Count: %u", pnanovdb_tree_get_tile_count_upper(buf, tree));
-                    ImGui::Text("Root Tile Count: %u", pnanovdb_tree_get_tile_count_root(buf, tree));
-                }
-
-                if (ImGui::CollapsingHeader("Root Information", ImGuiTreeNodeFlags_DefaultOpen))
-                {
-                    pnanovdb_tree_handle_t tree = pnanovdb_grid_get_tree(buf, grid);
-                    pnanovdb_root_handle_t root = pnanovdb_tree_get_root(buf, tree);
-
-                    pnanovdb_coord_t root_bbox_min = pnanovdb_root_get_bbox_min(buf, root);
-                    pnanovdb_coord_t root_bbox_max = pnanovdb_root_get_bbox_max(buf, root);
-                    ImGui::Text("Index BBox Min: (%d, %d, %d)", root_bbox_min.x, root_bbox_min.y, root_bbox_min.z);
-                    ImGui::Text("Index BBox Max: (%d, %d, %d)", root_bbox_max.x, root_bbox_max.y, root_bbox_max.z);
-
-                    ImGui::Text("Root Tile Count: %u", pnanovdb_root_get_tile_count(buf, root));
-                }
-
-                if (ImGui::CollapsingHeader("Blind Metadata"))
-                {
-                    uint32_t blindMetadataCount = pnanovdb_grid_get_blind_metadata_count(buf, grid);
-                    int64_t blindMetadataOffset = pnanovdb_grid_get_blind_metadata_offset(buf, grid);
-
-                    ImGui::Text("Blind Metadata Count: %u", blindMetadataCount);
-                    ImGui::Text("Blind Metadata Offset: %lld", (long long)blindMetadataOffset);
-
-                    if (blindMetadataCount > 0 && blindMetadataOffset != 0)
-                    {
-                        ImGui::Indent();
-                        for (uint32_t i = 0; i < blindMetadataCount && i < 10; ++i)
-                        { // Limit to 10 for UI
-                            pnanovdb_gridblindmetadata_handle_t metadata;
-                            metadata.address = pnanovdb_address_offset(grid.address, blindMetadataOffset + i * 288);
-
-                            uint32_t dataClass = pnanovdb_gridblindmetadata_get_data_class(buf, metadata);
-                            uint32_t dataType = pnanovdb_gridblindmetadata_get_data_type(buf, metadata);
-                            uint64_t valueCount = pnanovdb_gridblindmetadata_get_value_count(buf, metadata);
-                            uint32_t valueSize = pnanovdb_gridblindmetadata_get_value_size(buf, metadata);
-
-                            const char* className = "Unknown";
-                            switch (dataClass)
-                            {
-                            case 0:
-                                className = "Unknown";
-                                break;
-                            case 1:
-                                className = "Index Array";
-                                break;
-                            case 2:
-                                className = "Attribute Array";
-                                break;
-                            case 3:
-                                className = "Grid Name";
-                                break;
-                            case 4:
-                                className = "Channel Array";
-                                break;
-                            default:
-                                className = "Invalid";
-                                break;
-                            }
-
-                            ImGui::Text("Entry %u: %s (%u values, %u bytes each)", i, className, (unsigned)valueCount,
-                                        valueSize);
-                        }
-                        if (blindMetadataCount > 10)
-                        {
-                            ImGui::Text("... and %u more entries", blindMetadataCount - 10);
-                        }
-                        ImGui::Unindent();
-                    }
-                    else
-                    {
-                        ImGui::Text("No blind metadata present");
-                    }
-                }
-
-                if (ImGui::CollapsingHeader("Memory Usage"))
-                {
-                    ImGui::Text("Total Buffer Size: %llu bytes", (unsigned long long)(ptr->nanovdb_array->element_count *
-                                                                                      ptr->nanovdb_array->element_size));
-                    ImGui::Text("Element Size: %llu bytes", (unsigned long long)ptr->nanovdb_array->element_size);
-                    ImGui::Text("Element Count: %llu", (unsigned long long)ptr->nanovdb_array->element_count);
-                }
-            }
-            else
-            {
-                ImGui::Text("No NanoVDB file loaded");
-                ImGui::Text("Please load a NanoVDB file to view debug information");
-            }
-
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_code_editor)
-    {
-        pnanovdb_editor::CodeEditor::getInstance().setup(
-            &ptr->shader_name, &ptr->pending.update_shader, ptr->dialog_size, ptr->run_shader);
-        if (ImGui::Begin(
-                CODE_EDITOR, &ptr->window.show_code_editor, ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_MenuBar))
-        {
-            if (!pnanovdb_editor::CodeEditor::getInstance().render())
-            {
-                ptr->window.show_code_editor = false;
-            }
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_profiler)
-    {
-        if (ImGui::Begin(PROFILER, &ptr->window.show_profiler))
-        {
-            if (!pnanovdb_editor::Profiler::getInstance().render(&ptr->pending.update_memory_stats, delta_time))
-            {
-                ptr->window.show_profiler = false;
-            }
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-
-    if (ptr->window.show_console)
-    {
-        if (ImGui::Begin(CONSOLE, &ptr->window.show_console))
-        {
-            if (!pnanovdb_editor::Console::getInstance().render())
-            {
-                ptr->window.show_console = false;
-            }
-
-            ImGui::ProgressBar(ptr->progress.value, ImVec2(-1, 0), "");
-
-            ImVec2 pos = ImGui::GetItemRectMin();
-            ImVec2 size_bar = ImGui::GetItemRectSize();
-            ImVec2 text_size = ImGui::CalcTextSize(ptr->progress.text.c_str());
-            ImVec2 text_pos = ImVec2(pos.x + 5.f, pos.y + (size_bar.y - text_size.y) * 0.5f);
-
-            ImGui::GetWindowDrawList()->AddText(text_pos, ImGui::GetColorU32(ImGuiCol_Text), ptr->progress.text.c_str());
-
-            ImGui::End();
-        }
-        else
-        {
-            ImGui::End();
-        }
-    }
-}
-
-struct CameraBasisVectors
-{
-    pnanovdb_vec3_t right;
-    pnanovdb_vec3_t up;
-    pnanovdb_vec3_t forward;
-};
-
-static void calculateFrustumCorners(pnanovdb_camera_state_t& camera_state,
-                                    pnanovdb_camera_config_t& camera_config,
-                                    float aspectRatio,
-                                    pnanovdb_vec3_t corners[8],
-                                    CameraBasisVectors* basisVectors = nullptr,
-                                    float frustumScale = 1.f)
-{
-    pnanovdb_vec3_t eyePosition = pnanovdb_camera_get_eye_position_from_state(&camera_state);
-
-    // Calculate camera basis vectors
-    pnanovdb_vec3_t forward = camera_state.eye_direction;
-    pnanovdb_vec3_t up = camera_state.eye_up;
-    pnanovdb_vec3_t right = { forward.y * up.z - forward.z * up.y, forward.z * up.x - forward.x * up.z,
-                              forward.x * up.y - forward.y * up.x };
-
-    // Normalize right vector
-    float rightLength = sqrt(right.x * right.x + right.y * right.y + right.z * right.z);
-    if (rightLength > EPSILON)
-    {
-        right.x /= rightLength;
-        right.y /= rightLength;
-        right.z /= rightLength;
-    }
-    else
-    {
-        right = { 1.f, 0.f, 0.f };
-    }
-
-    up.x = right.y * forward.z - right.z * forward.y;
-    up.y = right.z * forward.x - right.x * forward.z;
-    up.z = right.x * forward.y - right.y * forward.x;
-
-    // Normalize the recalculated up vector
-    float upLength = sqrt(up.x * up.x + up.y * up.y + up.z * up.z);
-    if (upLength > EPSILON)
-    {
-        up.x /= upLength;
-        up.y /= upLength;
-        up.z /= upLength;
-    }
-    else
-    {
-        up = { 0.f, 1.f, 0.f };
-    }
-
-    if (basisVectors)
-    {
-        basisVectors->right = right;
-        basisVectors->up = up;
-        basisVectors->forward = forward;
-    }
-
-    // Handle reverse Z by swapping near/far planes
-    float nearPlane = camera_config.is_reverse_z ? camera_config.far_plane : camera_config.near_plane;
-    float farPlane = camera_config.is_reverse_z ? camera_config.near_plane : camera_config.far_plane;
-
-    nearPlane = std::max(nearPlane, EPSILON);
-    farPlane = std::min(farPlane, 10000000.f);
-
-    // Calculate frustum dimensions
-    float frustumNear = nearPlane * frustumScale;
-    float frustumFar = farPlane * frustumScale;
-
-    if (camera_config.is_orthographic)
-    {
-        frustumNear = 0.f; // place near plane at the camera position
-        frustumFar = std::max(frustumNear, frustumFar);
-    }
-
-    float nearHeight, nearWidth, farHeight, farWidth;
-
-    if (camera_config.is_orthographic)
-    {
-        float orthoHeight = camera_config.orthographic_y;
-        float orthoWidth = orthoHeight * aspectRatio;
-        nearHeight = farHeight = orthoHeight * frustumScale;
-        nearWidth = farWidth = orthoWidth * frustumScale;
-    }
-    else
-    {
-        float tanHalfFov = tan(camera_config.fov_angle_y * 0.5f);
-        nearHeight = frustumNear * tanHalfFov * frustumScale;
-        nearWidth = nearHeight * aspectRatio;
-        farHeight = frustumFar * tanHalfFov * frustumScale;
-        farWidth = farHeight * aspectRatio;
-    }
-
-    // Calculate near plane corners
-    pnanovdb_vec3_t nearCenter = { eyePosition.x + forward.x * frustumNear, eyePosition.y + forward.y * frustumNear,
-                                   eyePosition.z + forward.z * frustumNear };
-
-    corners[0] = { nearCenter.x - right.x * nearWidth * 0.5f - up.x * nearHeight * 0.5f,
-                   nearCenter.y - right.y * nearWidth * 0.5f - up.y * nearHeight * 0.5f,
-                   nearCenter.z - right.z * nearWidth * 0.5f - up.z * nearHeight * 0.5f };
-
-    corners[1] = { nearCenter.x + right.x * nearWidth * 0.5f - up.x * nearHeight * 0.5f,
-                   nearCenter.y + right.y * nearWidth * 0.5f - up.y * nearHeight * 0.5f,
-                   nearCenter.z + right.z * nearWidth * 0.5f - up.z * nearHeight * 0.5f };
-
-    corners[2] = { nearCenter.x + right.x * nearWidth * 0.5f + up.x * nearHeight * 0.5f,
-                   nearCenter.y + right.y * nearWidth * 0.5f + up.y * nearHeight * 0.5f,
-                   nearCenter.z + right.z * nearWidth * 0.5f + up.z * nearHeight * 0.5f };
-
-    corners[3] = { nearCenter.x - right.x * nearWidth * 0.5f + up.x * nearHeight * 0.5f,
-                   nearCenter.y - right.y * nearWidth * 0.5f + up.y * nearHeight * 0.5f,
-                   nearCenter.z - right.z * nearWidth * 0.5f + up.z * nearHeight * 0.5f };
-
-    // Calculate far plane corners
-    pnanovdb_vec3_t farCenter = { eyePosition.x + forward.x * frustumFar, eyePosition.y + forward.y * frustumFar,
-                                  eyePosition.z + forward.z * frustumFar };
-
-    corners[4] = { farCenter.x - right.x * farWidth * 0.5f - up.x * farHeight * 0.5f,
-                   farCenter.y - right.y * farWidth * 0.5f - up.y * farHeight * 0.5f,
-                   farCenter.z - right.z * farWidth * 0.5f - up.z * farHeight * 0.5f };
-
-    corners[5] = { farCenter.x + right.x * farWidth * 0.5f - up.x * farHeight * 0.5f,
-                   farCenter.y + right.y * farWidth * 0.5f - up.y * farHeight * 0.5f,
-                   farCenter.z + right.z * farWidth * 0.5f - up.z * farHeight * 0.5f };
-
-    corners[6] = { farCenter.x + right.x * farWidth * 0.5f + up.x * farHeight * 0.5f,
-                   farCenter.y + right.y * farWidth * 0.5f + up.y * farHeight * 0.5f,
-                   farCenter.z + right.z * farWidth * 0.5f + up.z * farHeight * 0.5f };
-
-    corners[7] = { farCenter.x - right.x * farWidth * 0.5f + up.x * farHeight * 0.5f,
-                   farCenter.y - right.y * farWidth * 0.5f + up.y * farHeight * 0.5f,
-                   farCenter.z - right.z * farWidth * 0.5f + up.z * farHeight * 0.5f };
-}
-
-static ImVec2 projectToScreen(const pnanovdb_vec3_t& worldPos,
-                              pnanovdb_camera_t* viewingCamera,
-                              float screenWidth,
-                              float screenHeight)
-{
-    pnanovdb_camera_mat_t view, proj;
-    pnanovdb_camera_get_view(viewingCamera, &view);
-    pnanovdb_camera_get_projection(viewingCamera, &proj, screenWidth, screenHeight);
-    pnanovdb_camera_mat_t viewProj = pnanovdb_camera_mat_mul(view, proj);
-
-    pnanovdb_vec4_t worldPos4 = { worldPos.x, worldPos.y, worldPos.z, 1.f };
-    pnanovdb_vec4_t clipPos = pnanovdb_camera_vec4_transform(worldPos4, viewProj);
-
-    if (clipPos.w > EPSILON)
-    {
-        float x = clipPos.x / clipPos.w;
-        float y = clipPos.y / clipPos.w;
-
-        // Convert from NDC to screen coordinates
-        float screenX = (x + 1.f) * 0.5f * screenWidth;
-        float screenY = (1.f - y) * 0.5f * screenHeight; // Flip Y coordinate
-
-        return ImVec2(screenX, screenY);
-    }
-
-    return ImVec2(std::numeric_limits<float>::quiet_NaN(), std::numeric_limits<float>::quiet_NaN());
-}
-
-static void drawCameraFrustum(
-    Instance* ptr, ImVec2 windowPos, ImVec2 windowSize, pnanovdb_camera_view_t& camera, int cameraIdx, float alpha = 1.f)
-{
-    ImDrawList* drawList = ImGui::GetWindowDrawList();
-
-    auto isPointFinite = [](const ImVec2& p) { return std::isfinite(p.x) && std::isfinite(p.y); };
-    auto isPointInsideWindow = [&](const ImVec2& p)
-    {
-        return (p.x >= windowPos.x && p.x <= windowPos.x + windowSize.x && p.y >= windowPos.y &&
-                p.y <= windowPos.y + windowSize.y);
-    };
-    auto isValidScreenPoint = [&](const ImVec2& p) { return isPointFinite(p) && isPointInsideWindow(p); };
-
-    pnanovdb_camera_t viewingCamera = {};
-    viewingCamera.state = ptr->render_settings->camera_state;
-    viewingCamera.config = ptr->render_settings->camera_config;
-
-    // Use a temporary perspective projection for overlay when both are ortho,
-    // so the frustum box depth is visible on screen
-    if (viewingCamera.config.is_orthographic && camera.configs[cameraIdx].is_orthographic)
-    {
-        viewingCamera.config.is_orthographic = PNANOVDB_FALSE;
-        viewingCamera.config.fov_angle_y = 0.785398163f; // ~45 deg
-        viewingCamera.config.near_plane = 0.1f;
-        viewingCamera.config.far_plane = 10000.f;
-    }
-
-    float aspectRatio = camera.configs[cameraIdx].aspect_ratio;
-    aspectRatio = aspectRatio <= 0.f ? windowSize.x / windowSize.y : aspectRatio;
-    pnanovdb_vec3_t frustumCorners[8];
-    CameraBasisVectors basisVectors;
-    calculateFrustumCorners(camera.states[cameraIdx], camera.configs[cameraIdx], aspectRatio, frustumCorners,
-                            &basisVectors, camera.frustum_scale);
-
-    ImVec2 screenCorners[8];
-    for (int i = 0; i < 8; i++)
-    {
-        screenCorners[i] = projectToScreen(frustumCorners[i], &viewingCamera, windowSize.x, windowSize.y);
-        screenCorners[i].x += windowPos.x;
-        screenCorners[i].y += windowPos.y;
-    }
-
-    ImU32 lineColor = IM_COL32(ImU32(camera.frustum_color.x * 255.f), ImU32(camera.frustum_color.y * 255.f),
-                               ImU32(camera.frustum_color.z * 255.f), ImU32(alpha * 255.f));
-    float frustumLineThickness = camera.frustum_scale * camera.frustum_line_width;
-
-    auto drawEdge = [&](int a, int b)
-    {
-        if (screenCorners[a] == screenCorners[b] || !isValidScreenPoint(screenCorners[a]) ||
-            !isValidScreenPoint(screenCorners[b]))
-        {
-            return;
-        }
-        drawList->AddLine(screenCorners[a], screenCorners[b], lineColor, frustumLineThickness);
-    };
-
-    // Near plane
-    drawEdge(0, 1);
-    drawEdge(1, 2);
-    drawEdge(2, 3);
-    drawEdge(3, 0);
-
-    // Far plane
-    drawEdge(4, 5);
-    drawEdge(5, 6);
-    drawEdge(6, 7);
-    drawEdge(7, 4);
-
-    // Sides
-    drawEdge(0, 4);
-    drawEdge(1, 5);
-    drawEdge(2, 6);
-    drawEdge(3, 7);
-
-    pnanovdb_vec3_t eyePosition = pnanovdb_camera_get_eye_position_from_state(&camera.states[cameraIdx]);
-
-    ImVec2 cameraScreenPos = projectToScreen(eyePosition, &viewingCamera, windowSize.x, windowSize.y);
-    cameraScreenPos.x += windowPos.x;
-    cameraScreenPos.y += windowPos.y;
-
-    if (!isValidScreenPoint(cameraScreenPos))
-    {
-        return;
-    }
-
-    // XYZ axes in RGB
-    pnanovdb_vec3_t forward = basisVectors.forward;
-    pnanovdb_vec3_t up = basisVectors.up;
-    pnanovdb_vec3_t right = basisVectors.right;
-
-    pnanovdb_vec3_t xAxisEnd = { eyePosition.x + right.x * camera.axis_length,
-                                 eyePosition.y + right.y * camera.axis_length,
-                                 eyePosition.z + right.z * camera.axis_length };
-    pnanovdb_vec3_t yAxisEnd = { eyePosition.x + up.x * camera.axis_length, eyePosition.y + up.y * camera.axis_length,
-                                 eyePosition.z + up.z * camera.axis_length };
-    pnanovdb_vec3_t zAxisEnd = { eyePosition.x + forward.x * camera.axis_length,
-                                 eyePosition.y + forward.y * camera.axis_length,
-                                 eyePosition.z + forward.z * camera.axis_length };
-
-    ImVec2 xAxisScreenPos = projectToScreen(xAxisEnd, &viewingCamera, windowSize.x, windowSize.y);
-    ImVec2 yAxisScreenPos = projectToScreen(yAxisEnd, &viewingCamera, windowSize.x, windowSize.y);
-    ImVec2 zAxisScreenPos = projectToScreen(zAxisEnd, &viewingCamera, windowSize.x, windowSize.y);
-
-    xAxisScreenPos.x += windowPos.x;
-    xAxisScreenPos.y += windowPos.y;
-    yAxisScreenPos.x += windowPos.x;
-    yAxisScreenPos.y += windowPos.y;
-    zAxisScreenPos.x += windowPos.x;
-    zAxisScreenPos.y += windowPos.y;
-
-    auto drawAxis = [&](const ImVec2& from, const ImVec2& to, ImU32 color)
-    {
-        if (from != to && isValidScreenPoint(to))
-        {
-            drawList->AddLine(from, to, color, camera.axis_thickness);
-        }
-    };
-
-    drawAxis(cameraScreenPos, xAxisScreenPos, IM_COL32(255, 0, 0, 255)); // right
-    drawAxis(cameraScreenPos, yAxisScreenPos, IM_COL32(0, 255, 0, 255)); // up
-    drawAxis(cameraScreenPos, zAxisScreenPos, IM_COL32(0, 0, 255, 255)); // forward
-
-    // Camera position
-    if (isValidScreenPoint(cameraScreenPos))
-    {
-        float posRadius = 1.5f * camera.axis_thickness;
-        ImU32 posColor = IM_COL32(222, 220, 113, ImU32(alpha * 255));
-
-        drawList->AddCircleFilled(cameraScreenPos, posRadius, posColor);
-    }
-}
-
-static void drawCameraViews(Instance* ptr)
-{
-    if (!ptr->camera_views)
-    {
-        return;
-    }
-
-    // Get the main viewport (central docking area)
-    ImGuiViewport* mainViewport = ImGui::GetMainViewport();
-    ImVec2 viewportPos = mainViewport->Pos;
-    ImVec2 viewportSize = mainViewport->Size;
-
-    // Create an overlay window to draw in the main viewport area
-    ImGui::SetNextWindowPos(viewportPos);
-    ImGui::SetNextWindowSize(viewportSize);
-    ImGui::SetNextWindowBgAlpha(0.f);
-
-    ImGuiWindowFlags windowFlags = ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoResize | ImGuiWindowFlags_NoMove |
-                                   ImGuiWindowFlags_NoScrollbar | ImGuiWindowFlags_NoSavedSettings |
-                                   ImGuiWindowFlags_NoInputs | ImGuiWindowFlags_NoFocusOnAppearing |
-                                   ImGuiWindowFlags_NoBringToFrontOnFocus;
-
-    ImGui::Begin("CameraFrustumOverlay", nullptr, windowFlags);
-    {
-        ImVec2 windowPos = ImGui::GetWindowPos();
-        ImVec2 windowSize = ImGui::GetWindowSize();
-
-        for (const auto& cameraPair : *ptr->camera_views)
-        {
-            pnanovdb_camera_view_t* camera = cameraPair.second;
-            if (!camera)
-            {
-                continue;
-            }
-
-            if (camera->is_visible)
-            {
-                bool isViewSelected =
-                    (!ptr->selected_camera_frustum.empty() && ptr->selected_camera_frustum == cameraPair.first);
-                int selected = isViewSelected ? ptr->camera_frustum_index[ptr->selected_camera_frustum] : -1;
-                // first draw non-selected cameras with lower alpha
-                for (int i = 0; i < camera->num_cameras; i++)
-                {
-                    if (i == selected)
-                    {
-                        continue;
-                    }
-                    drawCameraFrustum(ptr, windowPos, windowSize, *camera, i, isViewSelected ? 0.8f : 1.f);
-                }
-                // than draw selected camera with even lower alpha
-                if (selected >= 0)
-                {
-                    drawCameraFrustum(ptr, windowPos, windowSize, *camera, selected, 0.4f);
-                }
-            }
-        }
-    }
-    ImGui::End();
+    showConsoleWindow(ptr);
 }
 
 void update(pnanovdb_imgui_instance_t* instance)
@@ -1856,7 +239,7 @@ void update(pnanovdb_imgui_instance_t* instance)
 
     initializeDocking();
 
-    drawCameraViews(ptr);
+    drawCameraFrustums(ptr);
 
     createMenu(ptr);
 
@@ -1928,6 +311,16 @@ void update(pnanovdb_imgui_instance_t* instance)
 
         ImGuiFileDialog::Instance()->OpenDialog(
             "SelectShaderDirectoryDlgKey", "Select Shader Directory", nullptr, config);
+    }
+    if (ptr->pending.find_raster_file)
+    {
+        ptr->pending.find_raster_file = false;
+
+        IGFD::FileDialogConfig config;
+        config.path = ".";
+
+        ImGuiFileDialog::Instance()->OpenDialog(
+            "OpenRasterFileDlgKey", "Open Gaussian File", "Gaussian Files (*.npy *.npz *.ply){.npy,.npz,.ply}", config);
     }
 
     if (ImGuiFileDialog::Instance()->IsOpened())
